sources: refactor mercurial source into module.
Extract the Mercurial source into `sources/_mercurial.py`.

LP: #1648906
