+++ snapcraft/tests/sources/test_sources.py
@@ -478,128 +478,6 @@ class TestGit(SourceTestCase):
         self.assertEqual(raised.message, expected_message)
 
 
-class TestMercurial(SourceTestCase):
-
-    def test_pull(self):
-        hg = sources.Mercurial('hg://my-source', 'source_dir')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'clone', 'hg://my-source', 'source_dir'])
-
-    def test_pull_branch(self):
-        hg = sources.Mercurial('hg://my-source', 'source_dir',
-                               source_branch='my-branch')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'clone', '-u', 'my-branch', 'hg://my-source',
-             'source_dir'])
-
-    def test_pull_tag(self):
-        hg = sources.Mercurial('hg://my-source', 'source_dir',
-                               source_tag='tag')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'clone', '-u', 'tag', 'hg://my-source',
-             'source_dir'])
-
-    def test_pull_commit(self):
-        hg = sources.Mercurial('hg://my-source', 'source_dir',
-                               source_commit='2')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'clone', '-u', '2', 'hg://my-source',
-             'source_dir'])
-
-    def test_pull_existing(self):
-        self.mock_path_exists.return_value = True
-
-        hg = sources.Mercurial('hg://my-source', 'source_dir')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'pull', 'hg://my-source'])
-
-    def test_pull_existing_with_tag(self):
-        self.mock_path_exists.return_value = True
-
-        hg = sources.Mercurial('hg://my-source', 'source_dir',
-                               source_tag='tag')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'pull', '-r', 'tag', 'hg://my-source'])
-
-    def test_pull_existing_with_commit(self):
-        self.mock_path_exists.return_value = True
-
-        hg = sources.Mercurial('hg://my-source', 'source_dir',
-                               source_commit='2')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'pull', '-r', '2', 'hg://my-source'])
-
-    def test_pull_existing_with_branch(self):
-        self.mock_path_exists.return_value = True
-
-        hg = sources.Mercurial('hg://my-source', 'source_dir',
-                               source_branch='my-branch')
-        hg.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['hg', 'pull', '-b', 'my-branch', 'hg://my-source'])
-
-    def test_init_with_source_branch_and_tag_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Mercurial,
-            'hg://mysource', 'source_dir', source_tag='tag',
-            source_branch='branch')
-
-        expected_message = (
-            'can\'t specify both source-tag and source-branch for a mercurial '
-            'source')
-        self.assertEqual(raised.message, expected_message)
-
-    def test_init_with_source_commit_and_tag_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Mercurial,
-            'hg://mysource', 'source_dir', source_commit='2',
-            source_tag='tag')
-
-        expected_message = (
-            'can\'t specify both source-tag and source-commit for a mercurial '
-            'source')
-        self.assertEqual(raised.message, expected_message)
-
-    def test_init_with_source_commit_and_branch_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Mercurial,
-            'hg://mysource', 'source_dir', source_commit='2',
-            source_branch='branch')
-
-        expected_message = (
-            'can\'t specify both source-branch and source-commit for '
-            'a mercurial source')
-        self.assertEqual(raised.message, expected_message)
-
-    def test_init_with_source_depth_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Mercurial,
-            'hg://mysource', 'source_dir', source_depth=2)
-
-        expected_message = (
-            'can\'t specify source-depth for a mercurial source')
-        self.assertEqual(raised.message, expected_message)
-
-
 class TestSubversion(SourceTestCase):
 
     def test_pull_remote(self):
